+++ src/io/readers/file_readers.rs
@@ -1,3 +1,3 @@
 pub mod parquet_reader;
 pub mod sql_reader;
-pub mod tdf_blobs;
+pub mod tdf_blob_reader;
